Fix: Templates/Portfolio wrapping the Portfolio inside a QueryClientProvider and offering a PortfolioWrapped in case you need to use your own QueryClientProvider.
