+++ src/Components/Templates/Portfolio/Portfolio.tsx
@@ -1,120 +1,20 @@
-import React, { useEffect, useState } from 'react';
-import { QueryFunctionContext, useInfiniteQuery } from 'react-query';
-import { useInView } from 'react-intersection-observer';
+import React from 'react';
+import { QueryClient, QueryClientProvider } from 'react-query';
 
-import Loader from '../../Atoms/Icon/icons/Loader';
-import Filter, { IFilterProps } from '../../Organisms/Filter/Filter';
-import { SortDirectionEnum } from '../../Organisms/DataTable/Common/types';
-import { StaticDataTable } from '../../Organisms/DataTable/StaticDataTable';
-import { IStaticDataTableProps } from '../../Organisms/DataTable/StaticDataTable/StaticDataTable';
-
-export interface IPortfolioProps<FilterType, PortfolioType, ResponseType, PaginationType = number> {
-  /**
-   * Converts data synchronously.
-   *
-   * Note:
-   * - If extra data are required to populate the portfolio, please use the getData handler to enrich it.
-   */
-  convertData: (response: ResponseType) => Array<PortfolioType>;
-  /**
-   * Description of the filter
-   */
-  filter?: Pick<
-    IFilterProps<FilterType>,
-    'advancedSearchItems' | 'disableTabOutside' | 'initialValues' | 'localization' | 'searchBarItems'
-  >;
-  /**
-   * Handles the request for additional data (filtered, sorted, paginated).
-   *
-   * Note:
-   * - signal is used for cancelling the request
-   *    (see https://react-query-v3.tanstack.com/guides/query-cancellation )
-   * - pageParam is used for pagination management.
-   *    undefined for the first page or whenever there is a filter, sort or direction change
-   */
-  getData: (
-    filter: Partial<FilterType> | undefined,
-    sort: keyof PortfolioType | undefined,
-    direction: SortDirectionEnum | undefined,
-    pageParam: PaginationType | undefined,
-    signal: AbortSignal | undefined,
-  ) => Promise<ResponseType>;
-  /* Provide pagination information passed to the getData request */
-  getNextPageParam: (result: ResponseType, allResults: Array<ResponseType>) => PaginationType | undefined;
-  /* Handles error */
-  handleError: (error: unknown) => void;
-  /**
-   * Description of the static data table
-   */
-  table: Pick<IStaticDataTableProps<PortfolioType>, 'columns' | 'extra'>;
-}
+import PortfolioWrapped, { IPortfolioProps } from './PortfolioWrapped';
 
 const Portfolio = <FilterType, PortfolioType, ResponseType, PaginationType>(
   props: IPortfolioProps<FilterType, PortfolioType, ResponseType, PaginationType>,
 ) => {
-  const { filter, convertData, getData, getNextPageParam, handleError, table } = props;
-
-  const { ref, inView } = useInView();
-
-  const [filterValues, setFilterValues] = useState<Partial<FilterType> | undefined>();
-  const [sort, setSort] = useState<keyof PortfolioType>();
-  const [direction, setDirection] = useState<SortDirectionEnum>();
-
-  const { data, error, isFetching, fetchNextPage } = useInfiniteQuery(
-    ['portfolio', filterValues, sort, direction],
-    async (context: QueryFunctionContext) => {
-      const { signal, pageParam } = context;
-      return getData(filterValues, sort, direction, pageParam, signal);
-    },
-    {
-      getNextPageParam: getNextPageParam,
-    },
-  );
-
-  const onSortChange = (sortField?: keyof PortfolioType | undefined, sortDirection?: SortDirectionEnum | undefined) => {
-    setSort(sortField);
-    setDirection(sortDirection);
-  };
-
-  const onFilterChange = (value: Partial<FilterType> | undefined): void => {
-    setFilterValues(value);
-  };
-
-  useEffect(() => {
-    if (inView) {
-      fetchNextPage();
-    }
-  }, [inView, data]);
-
-  useEffect(() => {
-    if (error) {
-      handleError(error);
-    }
-  }, [error]);
+  const queryClient = new QueryClient();
 
   return (
     <div className='gds-portfolio'>
-      {filter && (
-        <div className='gds-portfolio-filter'>
-          <Filter {...filter} onChange={onFilterChange} />
-        </div>
-      )}
-
-      <div className='gds-portfolio-table'>
-        <StaticDataTable<PortfolioType>
-          {...table}
-          data={data?.pages.map(convertData).flat()}
-          onSortChange={onSortChange}
-          loading={isFetching ? <Loader size='3x' /> : undefined}
-        />
-        <div ref={ref} style={{ height: '20px' }} />
-      </div>
+      <QueryClientProvider client={queryClient}>
+        <PortfolioWrapped<FilterType, PortfolioType, ResponseType, PaginationType> {...props} />
+      </QueryClientProvider>
     </div>
   );
-};
-
-Portfolio.defaultProps = {
-  filter: undefined,
 };
 
 export default Portfolio;
