+++ src/Components/Templates/Portfolio/index.ts
@@ -1,2 +1,3 @@
-export type { IPortfolioProps } from './Portfolio';
+export type { IPortfolioProps } from './PortfolioWrapped';
+export { default as PortfolioWrapped } from './PortfolioWrapped';
 export { default as Portfolio } from './Portfolio';
